requirements.txt
python-telegram-bot==20.6
psycopg2-binary==2.9.10
python-dotenv==1.0.0
aiopg==1.4.0
requests==2.31.0
